requirements.txt
# requirements.txt
# Python dependencies for E-commerce Support Chatbot

Flask==3.0.0
Flask-CORS==4.0.0
mysql-connector-python==8.2.0
google-generativeai==0.3.2
python-dotenv==1.0.0